ticketing/requirements.txt
APScheduler==3.10.4
async-timeout==4.0.3
bcrypt==4.2.1
blinker==1.9.0
casttube==0.2.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
gunicorn==20.0.4
idna==3.7
ifaddr==0.2.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
paramiko==3.5.0
PlexAPI==4.14.0
protobuf==5.29.2
PyChromecast==14.0.5
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.32.3
six==1.16.0
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.1.3
zeroconf==0.136.2
zipp==3.21.0
